Make cluster test have predictible output
When `CLUSTER` is run in verbose mode on a hypertable, it prints the
chunks that are being clustered. The ordering of these chunks differs
in PG12 compared to previous versions, which caused the 'cluster' test
to fail.

The chunk index mappings that decide which chunks to cluster is now
sorted on chunk OID to make the output predicitible across PostgreSQL
versions. This comes at a slight cost, however, but shouldn't be much
overhead if one considers that the actual clustering is a lot more
costly.

One could optionally do the sorting only in case "verbose" output is
requested, but this doesn't seem worthwhile as it requires additional
code logic to handle both cases. It can always be done later as an
optimization.