Add collation when comparing versions
When comparing versions in the telemetry module when processing the
HTTP response, a direct call of `texteq` is made, but without using a
collation. This generate unnecessary errors in the log and also cause
the telemetry job to abort.

This commit fixes that by using the "C" collation when comparing the
version strings.