Make query test versioned and add PG12 output
This change modifies the query test to use the test template
mechanism so that we can capture the plan differences introduced by
Postgres 12 pruning append nodes.

The PG12 output file also contains a plan change which seems to
result in a GroupAggregate being replace by a less efficient Sort
plus HashAggregate.  This new behavior is still being investigated,
but it is still a correct plan (just potentially suboptimal).